+++ client/src/data/store.json
@@ -1,0 +1,9 @@
+{
+  "audios": [{
+    "name": "Crusade.mp3",
+    "title": "Crusade (Heavy Industry) - Kevin MacLeod"
+  }, {
+    "name": "Onion_Capers.mp3",
+    "title": "Onion Capers - Kevin MacLeod"
+  }]
+}
